requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
chromadb
sqlglot